requirements.txt
groq==0.11.0
python-dotenv==1.0.1
httpx==0.27.2
